Move DisplayPreferencesManager.cs to Users namespace
